requirements/dev.txt
# `requirements.txt` for development.

# Include common.
-r common.txt

# Dev only.
pytest==2.8.2
pytest-cov==2.2.0

coveralls==1.1

requests-mock==0.7.0
pytest-mock==0.9.0
mock==1.0.1
